feat: `frappe.utils.orjson_dumps`
